Move, Shoot, Enemies, and Time Health?
Implemented basic player controls including shooting,
and a basic enemy which can both be damaged (and killed)
by the bullets as well as hurt the player.

Had the idea to potentiall make health linked directly
to the timer, so not only does it constantly count down but
when the player gets hit you lose a couple seconds.