Enemy AI, AStar Pathfinding
Updated the enemy to be able to seek for the player, and when
it does it will chase the player down until it loses sight and
finishes the current path. Also made it so that the level will
generate an AStar matrix based off of the empty tiles (TODO:
update to floor tiles when a graphic exists for them).